chore: prepare release 0.15.0 (#943)
This PR was created by Knope. Merging it will create a new release

### Breaking Changes

#### Don't run `Command` steps in shell

The `Command` step no longer attempts to run the command in a default
shell for the detected operating system.
This fixes a compatibility issue with Windows.

If this change doesn't work for your workflow, please open an issue
describing your need so we can fix it.

Notably, using `&&` in a command (as was the case for some default
workflows) will no longer work. Instead, split this
into multiple `Command` steps.

PR #919 closes issue #918. Thanks for reporting @alex-way!

---------

Co-authored-by: GitHub <github-actions@github.com>
Co-authored-by: Dylan Anthony <dbanty@users.noreply.github.com>