chore: prepare release 0.16.0 (#957)
This PR was created by Knope. Merging it will create a new release

### Breaking Changes

#### Don't delete changesets for prereleases

Previously, using `PrepareRelease` to create a prerelease (for example,
with `--prerelease-label`) would delete all
changesets, just like a full release. This was a bug, but the fix is a
breaking change if you were
relying on that behavior.

### Features

#### Add a `shell` variable for `Command` steps

You can now add `shell=true` to a `Command` step to run the command in
the current shell.
This lets you opt in to the pre-0.15.0 behavior.

```toml
[[workflows.steps]]
type = "Command"
command = "echo $AN_ENV_VAR"
shell = true
```

Co-authored-by: GitHub <github-actions@github.com>